Improve server auth and add user to post

+++ express-backend/routes/posts.js
@@ -37,6 +37,7 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
+      creator: req.userData.userId,
       imagePath: url + "/images/" + req.file.filename, // req.file.filename provided by multer
     });
     post.save().then((createdPost) => {
@@ -67,10 +68,15 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({ message: "update successful" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "update successful" });
+      } else {
+        res.status(401).json({ message: "Not authorized" });
+      }
     });
   }
 );
@@ -108,8 +114,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    res.status(200).json({ message: "Post deleted" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId  }).then((result) => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Post deleted" });
+    } else {
+      res.status(401).json({ message: "Not authorized" });
+    }
   });
 });
